+++ src/app/_components/buttons/close-button/close-button.component.scss
