+++ src/app/_components/buttons/copy-button/copy-button.component.scss
