+++ src/app/_components/info/info-controller/info-controller.component.scss
