+++ src/app/_components/info/info-header/info-header.component.scss
