+++ src/app/_components/info/info-menu/info-menu.component.scss
