+++ src/app/_components/info/info-model/info-model.component.scss
