+++ src/app/_components/info/info-package/info-package.component.ts
