+++ src/app/_components/info/info-policy/info-policy.component.scss
