+++ src/app/_components/info/info-system/info-system.component.scss
