+++ src/app/_components/info/info-view/info-view.component.scss
