+++ src/app/_components/search-list/search-controllers-list/search-controllers-list.component.scss
