+++ src/app/_components/search-list/search-fields-list/search-fields-list.component.scss
