+++ src/app/_components/search-list/search-mixed-list/search-mixed-list.component.scss
