+++ src/app/_components/search-list/search-models-list/search-models-list.component.scss
