+++ src/app/_components/search-list/search-views-list/search-views-list.component.ts
